# swagger/masterclass.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GRPC Gateway API",
    "version": "1.0",
    "contact": {
      "name": "Bogo Tsvetkov",
      "url": "https://www.linkedin.com/in/bogomil-cvetkov-766a37b5/",
      "email": "cvetkov.bogomil@gmail.com"
    }
  },
  "tags": [
    {
      "name": "UserService"
    },
    {
      "name": "AccountService"
    },
    {
      "name": "TransferService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/account": {
      "get": {
        "summary": "List all accounts of a logged user",
        "description": "List all accounts of a logged user",
        "operationId": "AccountService_ListAccounts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ListAccountRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "AccountService"
        ]
      },
      "post": {
        "summary": "Use this endpoint to create a new bank account in a particular currency for an user",
        "description": "Create a new bank account for a logged user",
        "operationId": "AccountService_CreateAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateAccountRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateAccountReq"
            }
          }
        ],
        "tags": [
          "AccountService"
        ]
      }
    },
    "/account/{id}": {
      "get": {
        "summary": "List logged user account",
        "description": "List logged user account",
        "operationId": "AccountService_GetAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetAccountRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "AccountService"
        ]
      }
    },
    "/transfer": {
      "post": {
        "summary": "Use this endpoint to transfer money between a logged user account and another account",
        "description": "Transfer Money Endpoint",
        "operationId": "TransferService_CreateTransfer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateTransferRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Transfer money between accounts",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTransferReq"
            }
          }
        ],
        "tags": [
          "TransferService"
        ]
      }
    },
    "/user": {
      "post": {
        "summary": "Use this endpoint to create new user in the system",
        "description": "Register User Ednpoint",
        "operationId": "UserService_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateUserRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Register user params",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserReq"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/user/login": {
      "post": {
        "summary": "Use this endpoint to generate short-lived access and long-lived refresh token",
        "description": "Login",
        "operationId": "UserService_LoginUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/LoginUserRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Login params",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginUserReq"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/user/refresh-token": {
      "post": {
        "summary": "Use this endpoint to issue a new access-token using the provided refresh token in /login route",
        "description": "Refresh access token",
        "operationId": "UserService_RefreshToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/RefreshTokenRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Refresh token params",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RefreshTokenReq"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/user/verify": {
      "post": {
        "summary": "Use this endpoint to verify newly registered emails",
        "description": "Verify user account",
        "operationId": "UserService_VerifyEmail",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/VerifyEmailRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyEmailReq"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    }
  },
  "definitions": {
    "Account": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "owner": {
          "type": "string",
          "format": "int64"
        },
        "balance": {
          "type": "string",
          "format": "int64"
        },
        "currency": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "description": "User bank account type",
      "title": "Account type"
    },
    "CreateAccountReq": {
      "type": "object",
      "properties": {
        "currency": {
          "type": "string"
        }
      },
      "title": "Create"
    },
    "CreateAccountRes": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/Account"
        }
      }
    },
    "CreateTransferReq": {
      "type": "object",
      "properties": {
        "fromAccountId": {
          "type": "string",
          "format": "int64"
        },
        "toAccountId": {
          "type": "string",
          "format": "int64"
        },
        "amount": {
          "type": "string",
          "format": "int64"
        },
        "currency": {
          "type": "string"
        }
      },
      "description": "Transfer money between accounts",
      "title": "Create New Transfer",
      "required": [
        "fromAccountId",
        "toAccountId",
        "amount",
        "currency"
      ]
    },
    "CreateTransferRes": {
      "type": "object",
      "properties": {
        "transfer": {
          "$ref": "#/definitions/Transfer"
        }
      }
    },
    "CreateUserReq": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "description": "Register user params",
      "title": "Create User",
      "required": [
        "username",
        "password",
        "fullName",
        "email"
      ]
    },
    "CreateUserRes": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "GetAccountRes": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/Account"
        }
      }
    },
    "ListAccountRes": {
      "type": "object",
      "properties": {
        "accounts": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Account"
          }
        }
      }
    },
    "LoginUserReq": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "description": "Login params",
      "title": "Login User",
      "required": [
        "email",
        "password"
      ]
    },
    "LoginUserRes": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "RefreshTokenReq": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "description": "Refresh token params",
      "title": "Refresh token",
      "required": [
        "token"
      ]
    },
    "RefreshTokenRes": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "Transfer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "fromAccountId": {
          "type": "string",
          "format": "int64"
        },
        "toAccountId": {
          "type": "string",
          "format": "int64"
        },
        "amount": {
          "type": "string",
          "format": "int64"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "description": "Transfer between two accounts",
      "title": "Transfer type"
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "passChangedAt": {
          "type": "string",
          "format": "date-time"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "description": "Platform user type",
      "title": "User type"
    },
    "VerifyEmailReq": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
      },
      "title": "Verify Email"
    },
    "VerifyEmailRes": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}